// FE.Core.FW/src/apis/models/toefl-challenge/ExamLocationRoomModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ExamLocationRoomModel = {
  id?: string;
  examId?: string;
  examScheduleId?: string;
  examLocationId?: string;
  examLocationName?: string | null;
  name?: string | null;
  code?: string | null;
  amount?: number;
  order?: number;
  status?: boolean;
};